# openapi-comparison/1_APIBlueprint/openapi.yaml
swagger: '2.0'
info:
  title: Book Management API
  version: ''
  description: API for managing **books, users, and reviews**.
host: 127.0.0.1:5000
basePath: /
schemes:
  - http
paths:
...
  /api/v1/book-reviews:
    get:
      responses:
        '200':
          description: OK
          headers: {}
          examples:
            application/json:
              data:
                - id: 1
                  rating: 5
                  comment: Great
                  book_id: 1
                  user_id: 1
              meta:
                timestamp: '2026-01-01T12:00:00'
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    rating:
                      type: number
                    comment:
                      type: string
                    book_id:
                      type: number
                    user_id:
                      type: number
              meta:
                type: object
                properties:
                  timestamp:
                    type: string
            example:
              data:
                - id: 1
                  rating: 5
                  comment: Great
                  book_id: 1
                  user_id: 1
              meta:
                timestamp: '2026-01-01T12:00:00'
      summary: Get list of reviews
      operationId: Get list of reviews
      description: ''
      tags:
        - Reviews
      parameters: []
      produces:
        - application/json
      consumes: []
    post:
      responses:
        '201':
          description: Created
          headers: {}
          examples:
            application/json:
              data:
                id: 1
                rating: 5
                comment: Great
                book_id: 1
                user_id: 1
              meta:
                timestamp: '2026-01-01T12:00:00'
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  id:
                    type: number
                  rating:
                    type: number
                  comment:
                    type: string
                  book_id:
                    type: number
                  user_id:
                    type: number
              meta:
                type: object
                properties:
                  timestamp:
                    type: string
            example:
              data:
                id: 1
                rating: 5
                comment: Great
                book_id: 1
                user_id: 1
              meta:
                timestamp: '2026-01-01T12:00:00'
      summary: Create review
      operationId: Create review
      description: ''
      tags:
        - Reviews
      parameters:
        - name: body
          in: body
          schema:
            type: object
            properties:
              rating:
                type: number
              comment:
                type: string
              book_id:
                type: number
              user_id:
                type: number
            example:
              rating: 5
              comment: Great
              book_id: 1
              user_id: 1
      produces:
        - application/json
      consumes:
        - application/json
  /api/v1/book-reviews/{review_id}:
    get:
      responses:
        '200':
          description: OK
          headers: {}
          examples:
            application/json:
              data:
                id: 1
                rating: 5
                comment: Great
                book_id: 1
                user_id: 1
              meta:
                timestamp: '2026-01-01T12:00:00'
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  id:
                    type: number
                  rating:
                    type: number
                  comment:
                    type: string
                  book_id:
                    type: number
                  user_id:
                    type: number
              meta:
                type: object
                properties:
                  timestamp:
                    type: string
            example:
              data:
                id: 1
                rating: 5
                comment: Great
                book_id: 1
                user_id: 1
              meta:
                timestamp: '2026-01-01T12:00:00'
        '404':
          description: Not Found
          headers: {}
          examples: {}
      summary: Get review by ID
      operationId: Get review by ID
      description: ''
      tags:
        - Reviews
      parameters:
        - &ref_2
          name: review_id
          in: path
          description: ID of the review
          required: true
          x-example: '1'
          type: number
      produces:
        - application/json
      consumes: []
    delete:
      responses:
        '204':
          description: No Content
          headers: {}
          examples: {}
        '404':
          description: Not Found
          headers: {}
          examples: {}
      summary: Delete review
      operationId: Delete review
      description: ''
      tags:
        - Reviews
      parameters:
        - *ref_2
      consumes: []
...
securityDefinitions: {}
...